MOVE: split data and student into two seperate model files
